requirements.txt
aiogram==3.13.1
mysql-connector-python
cryptography
asyncio
babel
python-dotenv